pyproject.toml
[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.build_meta"

[project]
name = "data-pipeline-api-gateway"
version = "1.0.0"
description = "Production REST API gateway for vehicle diagnostic pipeline data"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "alembic>=1.13",
    "sqlalchemy>=2.0",  # alembic dependency only; queries are raw psycopg2
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "pytest-mock>=3.14",
    "httpx>=0.27",       # ASGI test client
    "black>=24",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-redis>=4",
    "faker>=25",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing -q"
log_cli = false

[tool.coverage.run]
omit = ["tests/*", "migrations/*", "scripts/*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
